+++ old-ui/app/components/token-cell.js
@@ -17,7 +17,12 @@ TokenCell.prototype.render = function () {
 
   return (
     h('li.token-cell', {
-      style: { cursor: network === '1' ? 'pointer' : 'default' },
+      style: {
+        cursor: network === '1' ? 'pointer' : 'default',
+        borderBottom: props.isLastTokenCell ? 'none' : '1px solid #e2e2e2',
+        padding: '20px 0',
+        margin: '0 20px',
+      },
       onClick: this.view.bind(this, address, userAddress, network),
     }, [
 
@@ -27,17 +32,24 @@ TokenCell.prototype.render = function () {
         network,
       }),
 
-      h('h3', `${string || 0} ${symbol}`),
+      h('h3', {
+        style: {
+          fontFamily: 'Nunito Bold',
+          fontSize: '14px',
+        },
+      }, `${string || 0} ${symbol}`),
 
       h('span', { style: { flex: '1 0 auto' } }),
 
-      h('span.fa.fa-trash', {
+      h('span.trash', {
         style: { cursor: 'pointer' },
         onClick: (event) => {
           event.stopPropagation()
           this.props.removeToken({ address, symbol, string, network, userAddress })
         },
       }, ''),
+
+      h('hr'),
 
       /*
       h('button', {